fix bug for A[ii] initialization; get uniform random graph working

+++ ctf/graph_aux.h
@@ -19,6 +19,6 @@ uint64_t read_graph(int myid, int ntask, const char *fpath, uint64_t **edge);
 uint64_t read_graph_mpiio(int myid, int ntask, const char *fpath, uint64_t **edge, char ***led);
 void processedges(char **led, uint64_t ned, int myid, int ntasks, uint64_t **edge, uint64_t start, bool eweights, wht ** vals);
 // uint64_t read_metis(int myid, int ntask, const char *fpath, uint64_t **edge, char ***led, int * n, uint64_t * start, bool * eweights, wht ** vals);
-uint64_t read_metis(int myid, int ntask, const char *fpath, std::vector<std::pair<uint64_t, uint64_t> > &edges, int * n, bool * e_weights, std::vector<wht> &eweights);
+uint64_t read_metis(int myid, int ntask, const char *fpath, std::vector<std::pair<uint64_t, uint64_t> > &edges, int64_t * n, bool * e_weights, std::vector<wht> &eweights);
 #endif
